Move spotdl backend, fix return code processing
